% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Zillionis Mouse Lung Example Dataset}
\format{
SingleCellExperiment object:
\describe{
  \item{number of genes}{24,248}
  \item{number of cells}{800}
  \item{number of cell-types}{4}
}
}
\source{
<https://pubmed.ncbi.nlm.nih.gov/30979687/>
}
\usage{
exampleData
}
\description{
Single-Cell Transcriptomics of Human and Mouse Lung Cancers Reveals Conserved Myeloid Populations across Individuals and Species
}
\keyword{datasets}
