% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mca10X}
\alias{mca10X}
\title{Mouse Cell Atlas: 10X}
\format{
named list object
}
\source{
<https://github.com/SydneyBioX/CLUEY>
}
\usage{
mca10X
}
\description{
Knowledge base for Mouse Cell Atlas 10X
}
\keyword{datasets}
