% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcaFACS}
\alias{mcaFACS}
\title{Mouse Cell Atlas: FACS}
\format{
named list object
}
\source{
<https://github.com/SydneyBioX/CLUEY>
}
\usage{
mcaFACS
}
\description{
Knowledge base for Mouse Cell Atlas FACS
}
\keyword{datasets}
