% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CLUEY.R
\name{run_CLUEY}
\alias{run_CLUEY}
\title{run_CLUEY}
\usage{
run_CLUEY(
  rna,
  modalities = NULL,
  knowledgebase,
  clus_method = "spectral",
  cor_method = "spearman",
  prop_genes = 0.25,
  k_limit = 20,
  sub_k = 3,
  min_cells = 30,
  recursive = TRUE,
  n_epochs = 50
)
}
\arguments{
\item{rna}{A numeric matrix of normalised gene expression data where rows are
features and columns are cells.}

\item{modalities}{A list of additional modalities, such as ATAC-seq or CITE-seq,
where rows are the features and columns are cells.}

\item{knowledgebase}{knowledgebase generated from the function `generateKnowledgeBase`.}

\item{clus_method}{Clustering method used. Default is Spectral clustering. Other option is K-means clustering.}

\item{cor_method}{Correlation method used when determining the most similar cell-types
for each cluster. Options are "Spearman" or "Pearson". Default is Spearman correlation.}

\item{prop_genes}{Proportion of genes to use when calculating the correlation of the expression
profiles between each cluster and knowledgeBase cell-type. Default is 0.25.}

\item{k_limit}{The maximum number of clusters to estimate. Default is 20.}

\item{sub_k}{The maximum number of clusters to estimate when performing iterative clustering. Default is 3.}

\item{min_cells}{The minimum number of cells that a cluster should contain. Default is 20 cells.}

\item{recursive}{A boolean specifying whether to perform clustering recursively. Default is TRUE.}

\item{nEpochs}{Number of epochs when training the autoencoder.}
}
\value{
A list containing the optimal number of clusters `optimal_k` and
cluster assignments and correlation scores for each cluster can be in `predictions`.
}
\description{
Perform cluster estimation of single or multi-modal single-cell data
}
